+++ src/index.ts
@@ -1,2 +1,3 @@
 export * from './typed-fetch';
 export * from './typed-fetch/api-operations';
+export * from './upload';